refactor: remove the web image-build feature
Frontend
- drop the per-row "Build" button and the "+ New image build" row
- delete BuildModal and its build API client functions/bindings
- drop the now-unused .modal-overlay/.modal styles

API
- delete handlers::builds, builds::BuildTracker and the 5 build routes
- drop BuildStatus/BuildResponse/FinalizeBuildRequest wire types
- move the builder-VM helpers bootstrap shares into handlers::builder_vm
- remove finalize_template_disk and prune_version, dead without builds

Docs
- replace the guide's derived-image build section with a removal note
