+++ docs/20-guides/m2image-builder.md
@@ -62,21 +62,11 @@ Rocky URL을 install URL로 직접 설정하면 안 된다.
 다음 단계인 고정 HTTP/GitHub Release 게시과 빈 호스트 설치 절차는
 [M2Image 레지스트리 태스크](../30-tasks/task-m2image-registry.md)를 따른다.
 
-## 웹에서 파생 이미지 빌드
-
-설치된 템플릿을 소스로, 대시보드에서 직접 패키지를 설치/삭제하고 새
-템플릿(또는 같은 alias의 새 버전)으로 저장할 수 있다. 배포판을 처음부터
-부트스트랩하는 것(위 CLI 경로)과는 다른 기능이다 — docker나 root 권한
-없이, Firecracker microVM 자체를 빌드 환경으로 쓴다.
-
-1. Images 화면에서 설치된 이미지의 "빌드" 클릭 (또는 "+ 새 이미지 빌드")
-2. 부팅된 builder VM 콘솔에서 패키지 설치/삭제
-3. "같은 이미지 갱신" 또는 "새 이미지로 저장"(새 alias 입력)
-4. 저장하면 새 템플릿 버전이 즉시 VM 생성에 쓸 수 있게 등록됨
-
-커널/initrd는 소스 템플릿 것을 그대로 공유한다 — 웹 빌드는 rootfs(패키지)만
-바꾼다. 완전히 새로운 배포판을 처음 추가하는 것은 여전히 이 문서 위쪽의
-CLI 경로(`build-m2images.sh`)를 쓴다.
+> [!note]
+> 설치된 템플릿에 패키지를 얹어 파생 이미지를 굽던 웹 빌드
+> (Images 화면의 "빌드" · "+ 새 이미지 빌드" 버튼과
+> `POST /api/images/{alias}/build` 계열 엔드포인트)는 제거되었다.
+> 웹에서 이미지를 만드는 경로는 아래 **배포판 부트스트랩** 하나다.
 
 ## 웹에서 배포판 부트스트랩
