+++ firecrab-api/src/main.rs
@@ -1,6 +1,5 @@
 mod artifacts;
 mod bootstrap;
-mod builds;
 mod console;
 mod error;
 mod extract;
