--- firecrab-frontend/src/bindings/BuildStatus.ts
@@ -1,3 +1,0 @@
-// Mirrors firecrab_api_types::BuildStatus (snake_case wire shape).
-
-export type BuildStatus = "booting" | "ready" | "installing" | "finalizing" | "succeeded" | "failed";
